Add manage.py
A new script for managing the test package repositories and building the
binary test artefacts (rpm, srpm and orig tarballs) from those.
Previously, test packages were maintained as separate branches in this
test data repository. However, this cluttered the refs space of the main
repository and made managing the test package "repositories" cumbersome
and also error prone, as many packages were basically managed in the
same git repository. Multiple branches also caused problems for
downstream users as the main repository is basically used as a git
submodule of git-buildpackage repository.

In the new model test package repositories are "serialized" as patches
accompanied with the proper metadata to re-construct the git repository
when needed. The new manage.py script is a tool for creating and
serializing the test package repositories as well as building binary
artefacts from these repositories. The repositories can be "perfectly"
re-constructed, meaning that all tags and local branches are restored
with full history and all object hashes are preserved as well.

The data layout shall be such that serialized data of a each repository
resides in a separate <NAME>.data directory. When changes to the
repository are required, or it needs to be built, it can be
reconstructed into <NAME>.repo directory.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>